Convert User to an interface
